test:   Separate decode and load tests into separate files.
